portal/requirements.txt
Flask==3.0.0
Flask-Login==0.6.3
Flask-Migrate==4.0.5
Flask-Session==0.8.0
Flask-Talisman==1.1.0
Flask-WTF==1.2.1
Flask-CORS==4.0.0
SQLAlchemy==2.0.23
Alembic==1.12.1
ldap3==2.9.1
redis==5.0.1
python-dotenv==1.0.0
python-json-logger==2.0.7
bcrypt==4.1.1
psycopg2-binary==2.9.9
WTForms==3.1.1
prometheus-client==0.19.0
gunicorn==21.2.0

# Development
pytest==7.4.3
pytest-flask==1.3.0
pytest-cov==4.1.0
flake8==6.1.0

